+++ pybamm/parameters/parameter_values.py
@@ -294,6 +294,17 @@ class ParameterValues:
                 "use 'Current function [A]' instead. The cell capacity can be accessed "
                 "as 'Cell capacity [A.h]', and used to calculate current from C-rate."
             )
+        for param in values:
+            if "surface area density" in param:
+                raise ValueError(
+                    "Parameters involving 'surface area density' have been renamed to "
+                    "'surface area to volume ratio' ('{}' found)".format(param)
+                )
+            if "reaction rate" in param:
+                raise ValueError(
+                    "Parameters involving 'reaction rate' have been replaced with "
+                    "'exchange-current density' ('{}' found)".format(param)
+                )
 
     def process_model(self, unprocessed_model, inplace=True):
         """Assign parameter values to a model.